Make nnx the default jax model impl
