+++ tpu_commons/models/jax/model_loader.py
@@ -15,11 +15,11 @@ def _get_model_architecture(config: PretrainedConfig) -> nn.Module:
     # NOTE: Use inline imports here, otherwise the normal imports
     # would cause JAX init failure when using multi hosts with Ray.
 
-    impl = os.getenv("MODEL_IMPL_TYPE", "flax_nn").lower()
+    impl = os.getenv("MODEL_IMPL_TYPE", "flax_nnx").lower()
     if impl == "flax_nn":
-        from tpu_commons.models.jax.llama import LlamaForCausalLM
+        from tpu_commons.models.jax.llama_nn import LlamaForCausalLM
     elif impl == "flax_nnx":
-        from tpu_commons.models.jax.llama_nnx import LlamaForCausalLM
+        from tpu_commons.models.jax.llama import LlamaForCausalLM
     else:
         raise NotImplementedError("Unsupported MODEL_IMPL_TYPE")
 
@@ -149,7 +149,7 @@ def get_model(
     rng: jax.Array,
     mesh: Mesh,
 ) -> nn.Module:
-    impl = os.getenv("MODEL_IMPL_TYPE", "flax_nn").lower()
+    impl = os.getenv("MODEL_IMPL_TYPE", "flax_nnx").lower()
     if impl == "flax_nn":
         return get_nn_model(vllm_config, rng, mesh)
     if impl == "flax_nnx":
